server - Copy/requirements.txt
matplotlib==3.8.4
pandas==2.2.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1